requirements.txt
aiosmtpd==1.4.6
alembic==1.16.5
atpublic==6.0.1
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.1
dnspython==2.8.0
elastic-transport==9.1.0
elasticsearch==9.1.1
email-validator==2.3.0
Flask==2.3.3
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==7.1.0
requests==2.32.5
rq==2.6.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==2.3.8
WTForms==3.2.1
